+++ SQL/New_patches/2026-01-04-Update_Login_Setup_Link.sql
@@ -1,0 +1,5 @@
+-- Update the Setup Guide link in the Study Description on the login page
+-- See: https://github.com/aces/Loris/issues/7071
+UPDATE Config 
+SET Value = REPLACE(Value, 'https://github.com/aces/Loris/wiki/Setup', 'https://acesloris.readthedocs.io/en/latest') 
+WHERE ConfigID = (SELECT ID FROM ConfigSettings WHERE Name = 'StudyDescription');