+++ modules/dataquery/jsx/nextsteps.tsx
@@ -114,19 +114,22 @@ function NextSteps(props: {
   }
 
   const expandIcon = <i
-    style={{transform: 'scaleY(2)', fontSize: '2em'}}
+    style={{
+      fontSize: '2em',
+      rotate: expanded ? '270deg' : '90deg',
+      color: '#064785',
+    }}
     className='fas fa-chevron-left'
-    onClick={() => setExpanded(!expanded)}
   ></i>;
   const style = expanded ? {
     background: 'white',
-    padding: '0.5em',
-    paddingLeft: '2em',
+    padding: '1rem 2rem',
+    width: '100%',
   } : {
-    display: 'none',
-    visibility: 'hidden' as const,
-    padding: '0.5em',
-    paddingLeft: '2em',
+    display: 'hidden',
+    padding: '0rem 2rem',
+    height: '0px',
+    width: '100%',
   };
 
   return (
@@ -134,30 +137,35 @@ function NextSteps(props: {
       position: 'fixed',
       bottom: 0,
       right: 0,
-      borderWidth: 'thin',
-      borderStyle: 'solid',
-      borderColor: 'black',
+      border: '1px solid black',
       // Fix the height size so it doesn't move when
       // expanded or collapsed
-      height: 120,
       // Make sure we're on top of the footer
       zIndex: 300,
+      backgroundColor: '#fff',
     }}>
+      <div
+        style={{display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'center',
+          gap: '16px',
+          padding: expanded ? '16px 3rem 8px' : '16px 3rem',
+          cursor: 'pointer',
+        }}
+        onClick={() => setExpanded(!expanded)}
+      >
+        <h3 style={{margin: '0'}}>{t('Next Steps', {ns: 'dataquery'})}</h3>
+        {expandIcon}</div>
       <div style={{
         display: 'flex',
         alignItems: 'stretch',
-        height: 120,
-        paddingRight: '14px',
       }}>
         <div style={style}>
-          <h3>{t('Next Steps', {ns: 'dataquery'})}</h3>
-          <div style={{display: 'flex'}}>
+          <div style={{display: 'flex', justifyContent: 'center', gap: '1rem'}}>
             {steps}
           </div>
         </div>
-        <div
-          style={{alignSelf: 'center'}}
-        >{expandIcon}</div>
+
       </div>
     </div>
   );