+++ CHANGELOG.md
@@ -12,6 +12,8 @@ changes in the following format: PR #1234***
 ### Core
 #### Summary
 - Minor release after v27 release
+#### Updates and Improvements
+- [docs] Update login page Setup Guide link to readthedocs (PR #7071)
 ### Notes For Existing Projects
 
 Upgrading LORIS requires following the upgrade process each major and minor release (bug fix releases can be script) to ensure the schema is up to date.